% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.apply.R
\name{array.apply}
\alias{array.apply}
\title{MoreParallelR::array.apply}
\usage{
array.apply(X, MARGIN, cores, FUN, ..., verbose = T)
}
\arguments{
\item{X}{An array.}

\item{MARGIN}{A vector giving the subscripts which the function will be applied over.}

\item{FUN}{The function to be applied.}

\item{...}{Optional arguments to \code{FUN}.}

\item{verbose}{Whether to print progress information.}
}
\value{
An array.
}
\description{
MoreParallelR::array.apply provides a convenient solution
for parallelizing the \code{apply} function on array. This
function first breaks the dimension specified by \code{MARGIN}
to a list of smaller arrays and then call the \code{mcapply}
function to achieve the rest of the parallelization.
}
\details{
To see better improvement by the parallelization, it is
preferred to have the runtime of \code{FUN} longer. In other
words, this solution works better when you have a heavy
workload in the function \code{FUN}.
}
\author{
Weiming Hu
}
